feat: 使用宏降低代码冗余
